version 2 of ui builder API
This version of the API seems a bit more ergonomic, while still exposing plenty of functionality for the end user. Probably a few improvements would make this even more robust. But for an initial version, I want to see what other's think and go from there. If the response is lackluster, I'll just make this for me, so I'll only add features that I am using actively.